Remove redundant com.aerospike.client.sdk.info.classes.IndexType.  Going forward, use com.aerospike.client.sdk.query.IndexType.

+++ client/src/main/java/com/aerospike/ael/visitor/VisitorUtils.java
@@ -77,7 +77,7 @@ public class VisitorUtils {
     }
 
     private final Map<Exp.Type, IndexType> expTypeToIndexType = Map.of(
-            Exp.Type.INT, IndexType.NUMERIC,
+            Exp.Type.INT, IndexType.INTEGER,
             Exp.Type.STRING, IndexType.STRING,
             Exp.Type.BLOB, IndexType.BLOB
     );
